+++ ChangeLog.md
@@ -4,6 +4,9 @@
 
 * Add digraph support under `M-k` and `/digraphs`
 * Add ECDSA-NIST256P-CHALLENGE support for Freenode via Tor
+* Load mask list on `/masks`
+* Add `C-x` to change to next network window
+* Allow `/clear NETWORK *` to clear all windows for the given network
 
 ## 2.17